+++ .gitignore
@@ -1,4 +1,8 @@
 /target
 bin
 */*fail*
+*fail
 /std
+*.rv
+*/target
+*/Cargo.lock
