+++ Cargo.toml
@@ -1,19 +1,16 @@
 [package]
 name = "quilt"
+authors = ["Nathaniel Fernandes <nathaniel.s.fernandes@gmail.com>"]
 version = "0.1.1"
 edition = "2021"
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
 [dependencies]
-peg = "0.8.1"
-ariadne = "0.1.5"
-fxhash = "0.2.1"
-colored = "2.0.0"
-paste = "1.0.11"
-arrayvec = "0.7.2"
-thiserror = "1.0.40"
-rand = "0.8.5"
+common       = { path = "./common" }
+parser       = { path = "./parser" }
+bytecode     = { path = "./bytecode" }
+compiler     = { path = "./compiler" }
+interpreter  = { path = "./interpreter" }
+disassembler = { path = "./disassembler" }
 
 [profile.release]
 opt-level = 3
