+++ common/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "common"
+authors = ["Nathaniel Fernandes <nathaniel.s.fernandes@gmail.com>"]
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+ariadne = "0.1.5"
+thiserror = "1.0.40"
+fxhash = "0.2.1"
