+++ interpreter/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "interpreter"
+authors = ["Nathaniel Fernandes <nathaniel.s.fernandes@gmail.com>"]
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+common = { path = "../common" }
+bytecode = { path = "../bytecode" }
+fxhash = "0.2.1"
+rand = "0.8.5"
+arrayvec = "0.7.2"
+paste = "1.0.12"
