+++ parser/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "parser"
+authors = ["Nathaniel Fernandes <nathaniel.s.fernandes@gmail.com>"]
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+peg = "0.8.1"
+common = { path = "../common" }
+fxhash = "0.2.1"
