--- src/backend/mod.rs
@@ -1,2 +1,0 @@
-pub mod utils;
-pub mod vm;
