--- src/backend/utils/mod.rs
@@ -1,2 +1,0 @@
-pub mod pool;
-pub mod rlevec;
