--- src/backend/vm/mod.rs
@@ -1,9 +1,0 @@
-pub mod arith;
-pub mod bytecode;
-pub mod chunk;
-pub mod compiler;
-pub mod dis;
-pub mod heap;
-pub mod tests;
-pub mod value;
-pub mod vm;
