--- src/frontend/mod.rs
@@ -1,3 +1,0 @@
-pub mod parser;
-pub use parser::*;
-pub mod debug;
