+++ src/lib.rs
@@ -1,29 +1,28 @@
-pub mod builtins;
-pub mod std;
+mod tests;
 
-pub mod backend;
-pub mod code;
-pub mod frontend;
-pub mod includes;
-pub mod shared;
-
-/// The prelude module contains all the functions and types that are available to use.
 pub mod prelude {
-    pub use crate::builtins::*;
+    pub use parser::includes::*;
+    pub use parser::node::Node;
+    pub use parser::{parse_code, parse_literal};
 
-    pub use crate::context_builtins;
-    pub use crate::generic_builtins;
-    pub use crate::specific_builtins;
+    pub use compiler::Compiler;
 
-    pub use paste::paste;
+    pub use interpreter::builtins::{Consumable, VmData};
+    pub use interpreter::qstd;
+    pub use interpreter::value::Value;
+    pub use interpreter::vm::VM;
+    pub use interpreter::Script;
 
-    pub use crate::code::*;
-    pub use crate::frontend::*;
-    pub use crate::includes::*;
-    pub use crate::shared::*;
-    pub use crate::std::*;
-    // pub use crate::treewalk::*;
-    pub use crate::backend::vm::value::*;
+    pub use interpreter::context_builtins;
+    pub use interpreter::generic_builtins;
+    pub use interpreter::specific_builtins;
 
-    pub use self::parser::parser::parse_code;
+    pub use common::error::{BuiltinError, Error, ErrorExt, ErrorS};
+    pub use common::sourcecache::SourceCache;
+
+    pub use disassembler::Disassembler;
+
+    pub mod builtins {
+        pub use interpreter::builtins::*;
+    }
 }
