+++ src/tests/closure.ql
